+++ zb-ucenter/zb-ucenter-rpc-service/src/main/java/pers/zb/ucenter/rpc/service/user/UserServiceImpl.java
@@ -1,6 +1,5 @@
 package pers.zb.ucenter.rpc.service.user;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
@@ -13,6 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 
+import pers.zb.common.util.AjaxResult;
 import pers.zb.common.util.Pager;
 import pers.zb.common.util.enums.FromSystemEnum;
 import pers.zb.common.util.util.CryptoUtils;
@@ -55,12 +55,32 @@ public class UserServiceImpl extends BaseServiceImpl<SysUser> implements UserSer
 
     @Transactional(propagation=Propagation.REQUIRED,rollbackFor=Exception.class)
     @Override
-    public void deleteUser(SysUser user) throws Exception{
+    public AjaxResult<String> deleteUser(Long userId) throws Exception{
+        AjaxResult<String> result = new AjaxResult<String>();
+        if (userId == null || "".equals(userId)) {
+            result.setCode(10001);
+            result.setMsg("userId为空");
+            return result;
+        }
+
+        SysUser user = userMapper.selectByPrimaryKey(userId);
+        if (user == null) {
+            result.setCode(10002);
+            result.setMsg("该用户不存在");
+            return result;
+        }
+        
+        if("admin".equals(user.getUserName())){//这里直接使用固定字符串比较了，真实项目不建议这种做法！扩展性不好。
+            result.setCode(10001);
+            result.setMsg("当前用户为管理员角色，不能删除");
+            return result;
+        }
         //删除用户
         userMapper.delete(user);
         
         //删除用户对于的角色
         userRoleService.deleteByUserId(user.getId());
+        return result;
     }
 
     @Transactional(propagation=Propagation.REQUIRED,rollbackFor=Exception.class)
@@ -97,7 +117,32 @@ public class UserServiceImpl extends BaseServiceImpl<SysUser> implements UserSer
     }
 
     @Override
-    public void deleteUsers(Long[] userIdArr) throws Exception {
+    public AjaxResult<String> deleteUsers(Long[] userIdArr) throws Exception {
+        AjaxResult<String> result = new AjaxResult<String>();
+        if (userIdArr == null || userIdArr.length <= 0) {
+            result.setCode(10001);
+            result.setMsg("userId为空");
+            return result;
+        }
+        
+        //验证选择的用户中是否有管理员角色的用户
+        Example exampleQuery = new Example(SysUser.class);
+        exampleQuery.createCriteria().andIn("id",Arrays.asList(userIdArr));
+        List<SysUser> userList = userMapper.selectByExample(exampleQuery);
+        
+        boolean bool = false;
+        for (SysUser sysUser : userList) {
+            if(sysUser != null && "admin".equals(sysUser.getUserName())){
+                bool = true;//存在管理员角色的用户
+                break;
+            }
+        }
+        if(bool){
+            result.setCode(10002);
+            result.setMsg("您不能删除拥有管理员角色的用户数据");
+            return result;
+        }
+        
         if(userIdArr != null && userIdArr.length > 0){
             for (int i = 0; i < userIdArr.length; i++) {
                 //删除用户
@@ -109,5 +154,6 @@ public class UserServiceImpl extends BaseServiceImpl<SysUser> implements UserSer
                 userRoleService.deleteByUserId(userIdArr[i]);
             }
         }
+        return result;
     }
 }
